Add test cases to cover low-hanging fruit
There are some cases here where I changed the actual code; most of those
instances are changing "return NotImplemented" or the like to a
docstring in abstract methods. Because it's impossible to run an
abstract method as-defined, it's not possible to test code coverage;
it's also not possible for the contents to matter at all, so why not
document what it should do?

The other two cases:
- added the attribute not found as an argument to AttributeError in
  busbus.entity.BaseEntity.__getattr__
- ProviderBase._new_entity is no longer an abstract method (it does
  nothing if you don't override it)