Inherit stop timezone from agency definition
"If omitted, the stop should be assumed to be located in the timezone
specified by agency_timezone in agency.txt."

There's some more language in the GTFS spec dealing with parent stations
that I need to look into as well.

Note that this also fixes the order GTFS files are read in, since
stops.txt now needs to be read after agency.txt.